Fix egregious duplication of vars in RnTypes
RnTypes contains a fairly intricate algorith to extract the
kind and type variables of an HsType. This algorithm carefully
maintains the separation between type variables and kind variables
so that the difference between -XPolyKinds and -XTypeInType can
be respected.

But it stupidly just concatenated the lists at the end. If a variable
were used as both a type and a kind, the algorithm would produce
*both*! This led to all kinds of problems, including #13988.

test case: ghci/scripts/T13988